+++ RELEASES.md
@@ -1,7 +1,8 @@
 Version 0.7.8 (2020-08-XX)
 ========================
  * New driver monitoring model
- * Hyundai Veloster 2019 support thanks to xps-gensis!
+ * Improved updater reliability and responsiveness
+ * Hyundai Genesis G70 2018 and Veloster 2019 support thanks to xps-genesis!
 
 Version 0.7.7 (2020-07-20)
 ========================